+++ src/containers/About.jsx
@@ -26,8 +26,8 @@ const About = React.createClass({
           <li><a href="https://facebook.github.io/react/">React</a></li>
           <li><a href="https://github.com/helior/react-highlighter">React Highlighter</a></li>
           <li><a href="https://github.com/rackt/react-router">React Router</a></li>
+          <li><a href="https://github.com/rackt/react-router-redux">React Router Redux</a></li>
           <li><a href="https://github.com/rackt/redux">Redux</a></li>
-          <li><a href="https://github.com/rackt/redux-simple-router">Redux Simple Router</a></li>
           <li><a href="https://github.com/gaearon/redux-thunk">Redux Thunk</a></li>
         </ul>
       </Paper>